@@ -1,29 +1,40 @@
 package com.bitvavo.api;
 
-import org.json.*;
+import java.io.IOException;
 import java.net.URI;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.TimeUnit;
+
 import javax.websocket.ClientEndpoint;
 import javax.websocket.CloseReason;
 import javax.websocket.ContainerProvider;
+import javax.websocket.DeploymentException;
 import javax.websocket.OnClose;
+import javax.websocket.OnError;
 import javax.websocket.OnMessage;
 import javax.websocket.OnOpen;
-import javax.websocket.OnError;
 import javax.websocket.Session;
 import javax.websocket.WebSocketContainer;
-import javax.websocket.PongMessage;
-import java.util.concurrent.TimeUnit;
-import java.util.*;
-import java.io.*;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
 
 @ClientEndpoint
 public class WebsocketClientEndpoint {
-    int reconnectTimer;
-    Bitvavo bitvavo;
-    Session userSession = null;
-    boolean restartWebsocket = true;
-    boolean keepBookCopy;
-    ArrayList<String> nonceArray;
+    private static final Logger LOGGER = LogManager.getLogger();
+
+    private int reconnectTimer;
+    private Bitvavo bitvavo;
+    private Session userSession = null;
+    private boolean restartWebsocket = true;
+    private boolean keepBookCopy;
     private MessageHandler messageHandler;
     private MessageHandler timeHandler;
     private MessageHandler marketsHandler;
@@ -50,147 +61,136 @@ public class WebsocketClientEndpoint {
     private MessageHandler withdrawalHistoryHandler;
     private MessageHandler authenticateHandler;
     private MessageHandler errorHandler;
-    HashMap<String, HashMap<String, MessageHandler>> subscriptionCandlesHandlerMap;
-    HashMap<String, MessageHandler> subscriptionAccountHandlerMap;
-    HashMap<String, MessageHandler> subscriptionTickerHandlerMap;
-    HashMap<String, MessageHandler> subscriptionTicker24hHandlerMap;
-    HashMap<String, MessageHandler> subscriptionTradesHandlerMap;
-    HashMap<String, MessageHandler> subscriptionBookUpdateHandlerMap;
-    HashMap<String, BookHandler> subscriptionBookHandlerMap;
+    private Map<String, HashMap<String, MessageHandler>> subscriptionCandlesHandlerMap;
+    private Map<String, MessageHandler> subscriptionAccountHandlerMap;
+    private Map<String, MessageHandler> subscriptionTickerHandlerMap;
+    private Map<String, MessageHandler> subscriptionTicker24hHandlerMap;
+    private Map<String, MessageHandler> subscriptionTradesHandlerMap;
+    private Map<String, MessageHandler> subscriptionBookUpdateHandlerMap;
+    private Map<String, BookHandler> subscriptionBookHandlerMap;
 
-    public WebsocketClientEndpoint(URI endpointURI, Bitvavo bitv) {
+    public WebsocketClientEndpoint(final URI endpointURI,final Bitvavo bitv) {
         try {
             bitvavo = bitv;
-            WebSocketContainer container = ContainerProvider.getWebSocketContainer();
+            final WebSocketContainer container = ContainerProvider.getWebSocketContainer();
             container.connectToServer(this, endpointURI);
-        } catch (Exception e) {
-            bitvavo.errorToConsole("Caught exception in instantiating websocket." + e);
-            this.reconnectTimer = 100;
+        } catch (final IOException | DeploymentException e) {
+            LOGGER.error("Caught exception in instantiating websocket.",e);
+            reconnectTimer = 100;
             retryConnecting(endpointURI);
         }
     }
 
-    public void retryConnecting(URI endpointURI) {
-        bitvavo.debugToConsole("Trying to reconnect.");
+    public void retryConnecting(final URI endpointURI) {
+        debug("Trying to reconnect.");
         try {
-            WebSocketContainer container = ContainerProvider.getWebSocketContainer();
+        	final WebSocketContainer container = ContainerProvider.getWebSocketContainer();
             container.connectToServer(this, endpointURI);
-        }
-        catch (javax.websocket.DeploymentException e) {
+        } catch (javax.websocket.DeploymentException e) {
             try {
                 TimeUnit.MILLISECONDS.sleep(this.reconnectTimer);
                 this.reconnectTimer = this.reconnectTimer * 2;
-                bitvavo.debugToConsole("We waited for " + this.reconnectTimer / 1000.0 + " seconds");
+                debug("We waited for " + this.reconnectTimer / 1000.0 + " seconds");
                 retryConnecting(endpointURI);
+            } catch(InterruptedException e1) {
+            	LOGGER.error(e1);                
             }
-            catch(InterruptedException err) {
-                return;
-            }
-        }
-        catch (Exception error) {
-            bitvavo.errorToConsole("unexpected exception caught");
+        } catch (final Exception error) {
+            LOGGER.error("unexpected exception caught",error);
             throw new RuntimeException(error);
         }
     }
 
     public void closeSocket() {
         try {
-            if (this.userSession != null) {
-                this.restartWebsocket = false;
-                this.userSession.close();
-                System.out.println(bitvavo.keepAliveThread);
-                bitvavo.keepAliveThread.interrupt();
+            if (userSession != null) {
+                restartWebsocket = false;
+                userSession.close();
+                LOGGER.trace(bitvavo.getKeepAliveThread());                
+                bitvavo.getKeepAliveThread().interrupt();
             }
         } catch (IOException e) {
-            e.printStackTrace();
+        	LOGGER.error(e);            
         }
     }
 
     @OnOpen
-    public void onOpen(Session userSession) {
-        bitvavo.debugToConsole("opening websocket");
+    public void onOpen(final Session userSession) {
+        debug("opening websocket");
         this.userSession = userSession;
         try {
             TimeUnit.MILLISECONDS.sleep(500);
+        } catch (final InterruptedException e) {
+            LOGGER.error(e);
         }
-        catch(InterruptedException ex) {
-            bitvavo.errorToConsole(ex.toString());
-        }
-        if(bitvavo.getApiKey() != "") {
-            long timestamp = System.currentTimeMillis();
-            JSONObject authenticate = new JSONObject();
+        if(!bitvavo.getApiKey().isEmpty()) {
+            final long timestamp = System.currentTimeMillis();
+            final JSONObject authenticate = new JSONObject();
             authenticate.put("action", "authenticate");
             authenticate.put("key", bitvavo.getApiKey());
             authenticate.put("signature", bitvavo.createSignature(timestamp, "GET", "/websocket", new JSONObject()));
             authenticate.put("timestamp", timestamp);
-            authenticate.put("window", Integer.toString(bitvavo.window));
-            this.sendMessage(authenticate.toString());
+            authenticate.put("window", Integer.toString(bitvavo.getWindow()));
+            sendMessage(authenticate.toString());
         }
-        if(bitvavo.activatedSubscriptionTicker) {
-            Iterator<String> markets = bitvavo.optionsSubscriptionTicker.keys();
-            while(markets.hasNext()) {
-                String market = markets.next();
-                this.sendMessage(bitvavo.optionsSubscriptionTicker.get(market).toString());
+        if (bitvavo.isActivatedSubscriptionTicker()) {
+            final Iterator<String> markets = bitvavo.getOptionsSubscriptionTicker().keys();
+            while (markets.hasNext()) {
+                sendMessage(bitvavo.getOptionsSubscriptionTicker().get(markets.next()).toString());
             }
         }
-        if(bitvavo.activatedSubscriptionTicker24h) {
-            Iterator<String> markets = bitvavo.optionsSubscriptionTicker24h.keys();
-            while(markets.hasNext()) {
-                String market = markets.next();
-                this.sendMessage(bitvavo.optionsSubscriptionTicker24h.get(market).toString());
+        if (bitvavo.isActivatedSubscriptionTicker24h()) {
+            final Iterator<String> markets = bitvavo.getOptionsSubscriptionTicker24h().keys();
+            while (markets.hasNext()) {                
+                sendMessage(bitvavo.getOptionsSubscriptionTicker24h().get(markets.next()).toString());
             }
         }
         // Account uses a threaded function, since we need a response on authenticate before we can send.
-        if(bitvavo.activatedSubscriptionAccount) {
-            WebsocketSendThread websocketSendThread = new WebsocketSendThread(bitvavo.optionsSubscriptionAccount, bitvavo, this);
+        if (bitvavo.isActivatedSubscriptionAccount()) {
+            final WebsocketSendThread websocketSendThread = new WebsocketSendThread(bitvavo.getOptionsSubscriptionAccount(), bitvavo, this);
             websocketSendThread.start();
         }
-        if(bitvavo.activatedSubscriptionCandles) {
-            Iterator<String> markets = bitvavo.optionsSubscriptionCandles.keys();
-            while(markets.hasNext()) {
-                String market = markets.next();
-                JSONObject intervalObject = bitvavo.optionsSubscriptionCandles.getJSONObject(market);
-                Iterator<String> intervals = intervalObject.keys();
-                while(intervals.hasNext()) {
-                    String interval = intervals.next();
-                    this.sendMessage(bitvavo.optionsSubscriptionCandles.getJSONObject(market).get(interval).toString());
+        if (bitvavo.isActivatedSubscriptionCandles()) {
+            final Iterator<String> markets = bitvavo.getOptionsSubscriptionCandles().keys();
+            while (markets.hasNext()) {
+                final String market = markets.next();
+                final JSONObject intervalObject = bitvavo.getOptionsSubscriptionCandles().getJSONObject(market);
+                final Iterator<String> intervals = intervalObject.keys();
+                while (intervals.hasNext()) {                    
+                    sendMessage(bitvavo.getOptionsSubscriptionCandles().getJSONObject(market).get(intervals.next()).toString());
                 }
             }
         }
-        if(bitvavo.activatedSubscriptionTrades) {
-            Iterator<String> markets = bitvavo.optionsSubscriptionTrades.keys();
-            while(markets.hasNext()) {
-                String market = markets.next();
-                this.sendMessage(bitvavo.optionsSubscriptionTrades.get(market).toString());
+        if (bitvavo.isActivatedSubscriptionTrades()) {
+            final Iterator<String> markets = bitvavo.getOptionsSubscriptionTrades().keys();
+            while (markets.hasNext()) {            	
+                sendMessage(bitvavo.getOptionsSubscriptionTrades().get(markets.next()).toString());
             }
         }
-        if(bitvavo.activatedSubscriptionBookUpdate) {
-            Iterator<String> markets = bitvavo.optionsSubscriptionBookUpdate.keys();
-            while(markets.hasNext()) {
-                String market = markets.next();
-                this.sendMessage(bitvavo.optionsSubscriptionBookUpdate.get(market).toString());
+        if (bitvavo.isActivatedSubscriptionBookUpdate()) {
+        	final Iterator<String> markets = bitvavo.getOptionsSubscriptionBookUpdate().keys();
+            while (markets.hasNext()) {
+                sendMessage(bitvavo.getOptionsSubscriptionBookUpdate().get(markets.next()).toString());
             }
         }
-        if(bitvavo.activatedSubscriptionBook) {
-            Iterator<String> markets = bitvavo.optionsSubscriptionBookFirst.keys();
-            while(markets.hasNext()) {
-                String market = markets.next();
-                this.sendMessage(bitvavo.optionsSubscriptionBookFirst.get(market).toString());
-                this.sendMessage(bitvavo.optionsSubscriptionBookSecond.get(market).toString());
+        if (bitvavo.isActivatedSubscriptionBook()) {
+            final Iterator<String> markets = bitvavo.getOptionsSubscriptionBookFirst().keys();
+            while (markets.hasNext()) {
+            	final String market = markets.next();
+                sendMessage(bitvavo.getOptionsSubscriptionBookFirst().get(market).toString());
+                sendMessage(bitvavo.getOptionsSubscriptionBookSecond().get(market).toString());
             }
         }
-        bitvavo.debugToConsole("We completed onOpen");
+        debug("We completed onOpen");
     }
 
     @OnError
-    public void onError(Session userSession, Throwable error) {
-        bitvavo.debugToConsole("We encountered an error: " + error);
-        error.printStackTrace();
-
+    public void onError(final Session userSession,final Throwable error) {
+        LOGGER.error("We encountered an error",error);
     }
 
-    private void copyHandlers(WebsocketClientEndpoint oldCE, WebsocketClientEndpoint newCE) {
-        if(oldCE.keepBookCopy) {
+    private void copyHandlers(final WebsocketClientEndpoint oldCE,final WebsocketClientEndpoint newCE) {
+        if (oldCE.keepBookCopy) {
             newCE.keepBookCopy = true;
         }
         newCE.copySubscriptionTickerHandler(oldCE.subscriptionTickerHandlerMap);
@@ -203,43 +203,42 @@ public class WebsocketClientEndpoint {
     }
 
     @OnClose
-    public void onClose(Session userSession, CloseReason reason) throws IOException {
-        bitvavo.debugToConsole("closing websocket " + reason);
+    public void onClose(final Session userSession,final CloseReason reason) {
+        debug("closing websocket " + reason);
         this.userSession = null;
-        if(bitvavo.getRemainingLimit() > 0 && this.restartWebsocket) {
+        if(bitvavo.getRemainingLimit() > 0 && restartWebsocket) {
             try {
-                bitvavo.authenticated = false;
-                WebsocketClientEndpoint clientEndPoint = new WebsocketClientEndpoint(new URI(bitvavo.wsUrl), bitvavo);
-                copyHandlers(bitvavo.ws, clientEndPoint);
+                bitvavo.setAuthenticated(false);
+                final WebsocketClientEndpoint clientEndPoint = new WebsocketClientEndpoint(new URI(bitvavo.getWsUrl()), bitvavo);
+                copyHandlers(bitvavo.getWs(), clientEndPoint);
                 clientEndPoint.addAuthenticateHandler(new WebsocketClientEndpoint.MessageHandler() {
                     public void handleMessage(JSONObject response) {
                         if(response.has("authenticated")) {
-                            bitvavo.authenticated = true;
-                            bitvavo.debugToConsole("We registered authenticated as true again");
+                            bitvavo.setAuthenticated(true);
+                            debug("We registered authenticated as true again");
                         }
                     }
                 });
                 clientEndPoint.addMessageHandler(new WebsocketClientEndpoint.MessageHandler() {
                     public void handleMessage(JSONObject response) {
-                        bitvavo.errorToConsole("Unexpected message: " + response);
+                        LOGGER.error("Unexpected message: {}",response);
                     }
                 });
-                bitvavo.ws = clientEndPoint;
-            }
-            catch(Exception ex) {
-                bitvavo.errorToConsole("We caught exception in reconnecting!" + ex);
+                bitvavo.setWs(clientEndPoint);
+            } catch (final Exception ex) {
+            	LOGGER.error("We caught exception in reconnecting!",ex);
             }
         } else {
-            if (this.restartWebsocket) {
-                bitvavo.debugToConsole("The websocket has been closed because your rate limit was reached, please wait till the ban is lifted and try again.");
+            if (restartWebsocket) {
+            	debug("The websocket has been closed because your rate limit was reached, please wait till the ban is lifted and try again.");
             } else {
-                bitvavo.debugToConsole("The websocket has been closed by the user.");
+            	debug("The websocket has been closed by the user.");
             }
         }
     }
 
-    public static Map<String, Object> jsonToMap(JSONObject json) throws JSONException {
-        Map<String, Object> returnMap = new HashMap<String, Object>();
+    public static Map<String, Object> jsonToMap(final JSONObject json) throws JSONException {
+        Map<String, Object> returnMap = new HashMap<>();
 
         if(json != JSONObject.NULL) {
             returnMap = toMap(json);
@@ -247,19 +246,17 @@ public class WebsocketClientEndpoint {
         return returnMap;
     }
 
-    public static Map<String, Object> toMap(JSONObject object) throws JSONException {
-        Map<String, Object> map = new HashMap<String, Object>();
+    public static Map<String, Object> toMap(final JSONObject object) throws JSONException {
+        Map<String, Object> map = new HashMap<>();
 
-        Iterator<String> keysItr = object.keys();
-        while(keysItr.hasNext()) {
-            String key = keysItr.next();
+        final Iterator<String> keysItr = object.keys();
+        while (keysItr.hasNext()) {
+            final String key = keysItr.next();
             Object value = object.get(key);
 
-            if(value instanceof JSONArray) {
+            if (value instanceof JSONArray) {
                 value = toList((JSONArray) value);
-            }
-
-            else if(value instanceof JSONObject) {
+            } else if(value instanceof JSONObject) {
                 value = toMap((JSONObject) value);
             }
             map.put(key, value);
@@ -267,15 +264,13 @@ public class WebsocketClientEndpoint {
         return map;
     }
 
-    public static List<Object> toList(JSONArray array) throws JSONException {
-        List<Object> list = new ArrayList<Object>();
+    public static List<Object> toList(final JSONArray array) throws JSONException {
+        final List<Object> list = new ArrayList<>();
         for(int i = 0; i < array.length(); i++) {
             Object value = array.get(i);
-            if(value instanceof JSONArray) {
+            if (value instanceof JSONArray) {
                 value = toList((JSONArray) value);
-            }
-
-            else if(value instanceof JSONObject) {
+            } else if (value instanceof JSONObject) {
                 value = toMap((JSONObject) value);
             }
             list.add(value);
@@ -283,26 +278,26 @@ public class WebsocketClientEndpoint {
         return list;
     }
 
-    private List<List<String>> sortAndInsert(List<List<String>> update, List<List<String>> book, boolean asksCompare) {
+    private List<List<String>> sortAndInsert(final List<List<String>> update,final List<List<String>> book,final boolean asksCompare) {
         for(int i = 0; i < update.size(); i++) {
             boolean updateSet = false;
-            List<String> updateEntry = update.get(i);
-            for(int j = 0; j < book.size(); j++) {
-                List<String> bookItem = book.get(j);
-                if(asksCompare) {
-                    if(Float.parseFloat(updateEntry.get(0)) < Float.parseFloat(bookItem.get(0))) {
+            final List<String> updateEntry = update.get(i);
+            for (int j = 0; j < book.size(); j++) {
+                final List<String> bookItem = book.get(j);
+                if (asksCompare) {
+                    if (Float.parseFloat(updateEntry.get(0)) < Float.parseFloat(bookItem.get(0))) {
                         book.add(j, updateEntry);
                         updateSet = true;
                         break;
                     }
                 } else {
-                    if(Float.parseFloat(updateEntry.get(0)) > Float.parseFloat(bookItem.get(0))) {
+                    if (Float.parseFloat(updateEntry.get(0)) > Float.parseFloat(bookItem.get(0))) {
                         book.add(j, updateEntry);
                         updateSet = true;
                         break;
                     }
                 }
-                if(Float.parseFloat(bookItem.get(0)) == Float.parseFloat(updateEntry.get(0))) {
+                if (Float.parseFloat(bookItem.get(0)) == Float.parseFloat(updateEntry.get(0))) {
                     if(Float.parseFloat(updateEntry.get(1)) > 0.0) {
                         book.set(j, updateEntry);
                         updateSet = true;
@@ -314,110 +309,109 @@ public class WebsocketClientEndpoint {
                     }
                 }
             }
-            if(updateSet == false) {
+            if (!updateSet) {
                 book.add(updateEntry);
             }
         }
         return book;
     }
 
+    private void debug(final String message) {
+    	if(bitvavo.isDebugging() && LOGGER.isDebugEnabled())  {
+          LOGGER.debug(message);
+        }
+    }
+    
     @SuppressWarnings("unchecked")
     @OnMessage
-    public void onMessage(String message) {
+    public void onMessage(final String message) {
         String market;
-        JSONObject response = new JSONObject(message);
-        bitvavo.debugToConsole("FULLRESPONSE: " + response);
-        if(response.has("error")) {
+        final JSONObject response = new JSONObject(message);
+        debug("FULLRESPONSE: " + response);
+        if (response.has("error")) {
             bitvavo.errorRateLimit(response);
-            if(this.errorHandler != null) {
-                this.errorHandler.handleMessage(response);
+            if(errorHandler != null) {
+                errorHandler.handleMessage(response);
                 return;
             }
         }
-        if(response.has("event")) {
+        if (response.has("event")) {
             if(response.getString("event").equals("subscribed")) {
-                JSONObject channel = response.getJSONObject("subscriptions");
+                final JSONObject channel = response.getJSONObject("subscriptions");
                 String subscribedString = "We are now subscribed to the following channels: ";
-                Iterator<String> keys = channel.keys();
+                final Iterator<String> keys = channel.keys();
                 while(keys.hasNext()) {
                     String key = keys.next();
                     subscribedString = subscribedString + key + ", ";
                 }
-                bitvavo.debugToConsole(subscribedString.substring(0, subscribedString.length() - 2));
-            }
-            else if(response.getString("event").equals("authenticate")) {
-                if (this.authenticateHandler != null) {
-                    this.authenticateHandler.handleMessage(response);
+                debug(subscribedString.substring(0, subscribedString.length() - 2));
+            } else if(response.getString("event").equals("authenticate")) {
+                if (authenticateHandler != null) {
+                    authenticateHandler.handleMessage(response);
                 }
-            }
-            else if(response.getString("event").equals("trade")) {
+            } else if(response.getString("event").equals("trade")) {
                 market = response.getString("market");
-                if (this.subscriptionTradesHandlerMap != null) {
-                    if(this.subscriptionTradesHandlerMap.get(market) != null) {
-                        this.subscriptionTradesHandlerMap.get(market).handleMessage(response);
+                if (subscriptionTradesHandlerMap != null) {
+                    if (subscriptionTradesHandlerMap.get(market) != null) {
+                        subscriptionTradesHandlerMap.get(market).handleMessage(response);
                     }
                 }
-            }
-            else if (response.getString("event").equals("fill")) {
+            } else if (response.getString("event").equals("fill")) {
                 market = response.getString("market");
-                if (this.subscriptionAccountHandlerMap != null) {
-                    if(this.subscriptionAccountHandlerMap.get(market) != null) {
-                        this.subscriptionAccountHandlerMap.get(market).handleMessage(response);
+                if (subscriptionAccountHandlerMap != null) {
+                    if (subscriptionAccountHandlerMap.get(market) != null) {
+                        subscriptionAccountHandlerMap.get(market).handleMessage(response);
                     }
                 }
-            }
-            else if (response.getString("event").equals("order")) {
+            } else if (response.getString("event").equals("order")) {
                 market = response.getString("market");
-                if (this.subscriptionAccountHandlerMap != null) {
-                    if(this.subscriptionAccountHandlerMap.get(market) != null) {
-                        this.subscriptionAccountHandlerMap.get(market).handleMessage(response);
+                if (subscriptionAccountHandlerMap != null) {
+                    if (subscriptionAccountHandlerMap.get(market) != null) {
+                        subscriptionAccountHandlerMap.get(market).handleMessage(response);
                     }
                 }
-            }
-            else if (response.getString("event").equals("ticker")) {
+            } else if (response.getString("event").equals("ticker")) {
                 market = response.getString("market");
-                if (this.subscriptionTickerHandlerMap != null) {
-                    if(this.subscriptionTickerHandlerMap.get(market) != null) {
-                        this.subscriptionTickerHandlerMap.get(market).handleMessage(response);
+                if (subscriptionTickerHandlerMap != null) {
+                    if (subscriptionTickerHandlerMap.get(market) != null) {
+                        subscriptionTickerHandlerMap.get(market).handleMessage(response);
                     }
                 }
-            }
-            else if (response.getString("event").equals("ticker24h")) {
-                JSONArray data = response.getJSONArray("data");
-                if (this.subscriptionTicker24hHandlerMap != null) {
+            } else if (response.getString("event").equals("ticker24h")) {
+                final JSONArray data = response.getJSONArray("data");
+                if (subscriptionTicker24hHandlerMap != null) {
                     for (int i = 0; i < data.length(); i++) {
-                        JSONObject ticker = data.getJSONObject(i);
+                        final JSONObject ticker = data.getJSONObject(i);
                         market = ticker.getString("market");
-                        if (this.subscriptionTicker24hHandlerMap.get(market) != null) {
-                            this.subscriptionTicker24hHandlerMap.get(market).handleMessage(ticker);
+                        if (subscriptionTicker24hHandlerMap.get(market) != null) {
+                            subscriptionTicker24hHandlerMap.get(market).handleMessage(ticker);
                         }
                     }
                 }
-            }
-            else if (response.getString("event").equals("book")) {
+            } else if (response.getString("event").equals("book")) {
                 market = response.getString("market");
-                if (this.subscriptionBookUpdateHandlerMap != null) {
-                    if(this.subscriptionBookUpdateHandlerMap.get(market) != null) {
-                        this.subscriptionBookUpdateHandlerMap.get(market).handleMessage(response);
+                if (subscriptionBookUpdateHandlerMap != null) {
+                    if(subscriptionBookUpdateHandlerMap.get(market) != null) {
+                        subscriptionBookUpdateHandlerMap.get(market).handleMessage(response);
                     }
                 }
                 if (keepBookCopy) {
-                    if(this.subscriptionBookHandlerMap != null) {
-                        if(this.subscriptionBookHandlerMap.get(market) != null) {
-                            Map<String, Object> responseMap = jsonToMap(response);
+                    if(subscriptionBookHandlerMap != null) {
+                        if(subscriptionBookHandlerMap.get(market) != null) {
+                            final Map<String, Object> responseMap = jsonToMap(response);
                             market = (String)responseMap.get("market");
                             boolean restartLocalBook = false;
 
-                            Map<String, Object> bidsAsks = (Map<String, Object>)bitvavo.book.get(market);
+                            final Map<String, Object> bidsAsks = (Map<String, Object>)bitvavo.getBook().get(market);
                             List<List<String>> bids = (List<List<String>>)bidsAsks.get("bids");
                             List<List<String>> asks = (List<List<String>>)bidsAsks.get("asks");
                             
-                            List<List<String>> bidsInput = (List<List<String>>)responseMap.get("bids");
-                            List<List<String>> asksInput = (List<List<String>>)responseMap.get("asks");
+                            final List<List<String>> bidsInput = (List<List<String>>)responseMap.get("bids");
+                            final List<List<String>> asksInput = (List<List<String>>)responseMap.get("asks");
 
                             if( (int)responseMap.get("nonce") != Integer.parseInt((String)bidsAsks.get("nonce")) + 1) {
                                 restartLocalBook = true;
-                                bitvavo.websocketObject.subscriptionBook(market, this.subscriptionBookHandlerMap.get(market));
+                                bitvavo.getWebsocketObject().subscriptionBook(market, this.subscriptionBookHandlerMap.get(market));
                             }
                             if(!restartLocalBook) {
                                 bids = sortAndInsert(bidsInput, bids, false);
@@ -425,357 +419,332 @@ public class WebsocketClientEndpoint {
                                 bidsAsks.put("bids", bids);
                                 bidsAsks.put("asks", asks);
                                 bidsAsks.put("nonce", Integer.toString((int)responseMap.get("nonce")));
-                                bitvavo.book.put(market, bidsAsks);
+                                bitvavo.getBook().put(market, bidsAsks);
                             
-                                this.subscriptionBookHandlerMap.get(market).handleBook((Map<String, Object>)bitvavo.book.get(market));
+                                subscriptionBookHandlerMap.get(market).handleBook((Map<String, Object>)bitvavo.getBook().get(market));
                             }
                         }
                     }
                 }
-            }
-            else if (response.getString("event").equals("candle")) {
+            } else if (response.getString("event").equals("candle")) {
                 market = response.getString("market");
-                String interval = response.getString("interval");
-                if (this.subscriptionCandlesHandlerMap.get(market) != null) {
-                    if(this.subscriptionCandlesHandlerMap.get(market).get(interval) != null) {
-                        this.subscriptionCandlesHandlerMap.get(market).get(interval).handleMessage(response);
+                final String interval = response.getString("interval");
+                if (subscriptionCandlesHandlerMap.get(market) != null) {
+                    if(subscriptionCandlesHandlerMap.get(market).get(interval) != null) {
+                        subscriptionCandlesHandlerMap.get(market).get(interval).handleMessage(response);
                     }
                 }
             }
-        }
-        else if(response.has("action")) {
-            if(response.getString("action").equals("getTime")) {
-                if (this.timeHandler != null) {
-                    this.timeHandler.handleMessage(response);
+        } else if(response.has("action")) {
+            if (response.getString("action").equals("getTime")) {
+                if (timeHandler != null) {
+                    timeHandler.handleMessage(response);
                 }
-            }
-            else if(response.getString("action").equals("getMarkets")) {
-                if (this.marketsHandler != null) {
-                    this.marketsHandler.handleMessage(response);
+            } else if(response.getString("action").equals("getMarkets")) {
+                if (marketsHandler != null) {
+                    marketsHandler.handleMessage(response);
                 }
-            }
-            else if(response.getString("action").equals("getAssets")) {
-                if (this.assetsHandler != null) {
-                    this.assetsHandler.handleMessage(response);
+            } else if(response.getString("action").equals("getAssets")) {
+                if (assetsHandler != null) {
+                    assetsHandler.handleMessage(response);
                 }
-            }
-            else if(response.getString("action").equals("getBook")) {
-                if (this.bookHandler != null) {
-                    this.bookHandler.handleMessage(response);
+            } else if(response.getString("action").equals("getBook")) {
+                if (bookHandler != null) {
+                    bookHandler.handleMessage(response);
                 }
                 if (keepBookCopy) {
                     if(this.subscriptionBookHandlerMap != null) {
                         market = response.getJSONObject("response").getString("market");
-                        if(this.subscriptionBookHandlerMap.get(market) != null) {
-                            Map<String, Object> bidsAsks = (Map<String, Object>)bitvavo.book.get(market);
-                            List<List<String>> bids = (List<List<String>>)bidsAsks.get("bids");
-                            List<List<String>> asks = (List<List<String>>)bidsAsks.get("asks");
+                        if(subscriptionBookHandlerMap.get(market) != null) {
+                            final Map<String, Object> bidsAsks = (Map<String, Object>)bitvavo.getBook().get(market);
+                            //final List<List<String>> bids = (List<List<String>>)bidsAsks.get("bids");
+                            //final List<List<String>> asks = (List<List<String>>)bidsAsks.get("asks");
 
-                            Map<String, Object> bookentry = jsonToMap(response.getJSONObject("response"));
+                            final Map<String, Object> bookentry = jsonToMap(response.getJSONObject("response"));
 
-                            List<List<String>> bidsInput = (List<List<String>>)bookentry.get("bids");
-                            List<List<String>> asksInput = (List<List<String>>)bookentry.get("asks");
+                            final List<List<String>> bidsInput = (List<List<String>>)bookentry.get("bids");
+                            final List<List<String>> asksInput = (List<List<String>>)bookentry.get("asks");
 
                             bidsAsks.put("bids", bidsInput);
                             bidsAsks.put("asks", asksInput);
                             bidsAsks.put("nonce", Integer.toString(response.getJSONObject("response").getInt("nonce")));
-                            bitvavo.book.put(market, bidsAsks);
+                            bitvavo.getBook().put(market, bidsAsks);
                         
-                            this.subscriptionBookHandlerMap.get(market).handleBook((Map<String, Object>)bitvavo.book.get(market));
+                            subscriptionBookHandlerMap.get(market).handleBook((Map<String, Object>)bitvavo.getBook().get(market));
                         }
                     }
                 }
-            }
-            else if(response.getString("action").equals("getTrades")) {
-                if (this.tradesHandler != null) {
-                    this.tradesHandler.handleMessage(response);
+            } else if(response.getString("action").equals("getTrades")) {
+                if (tradesHandler != null) {
+                    tradesHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("getCandles")) {
+                if (candlesHandler != null) {
+                    candlesHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("getTicker24h")) {
+                if (ticker24hHandler != null) {
+                    ticker24hHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("getTickerPrice")) {
+                if (tickerPriceHandler != null) {
+                    tickerPriceHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("getTickerBook")) {
+                if (tickerBookHandler != null) {
+                    tickerBookHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateCreateOrder")) {
+                if (createOrderHandler != null) {
+                    createOrderHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateGetOrder")) {
+                if (getOrderHandler != null) {
+                    getOrderHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateUpdateOrder")) {
+                if (updateOrderHandler != null) {
+                    updateOrderHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateCancelOrder")) {
+                if (cancelOrderHandler != null) {
+                    cancelOrderHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateGetOrders")) {
+                if (getOrdersHandler != null) {
+                    getOrdersHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateCancelOrders")) {
+                if (cancelOrdersHandler != null) {
+                    cancelOrdersHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateGetOrdersOpen")) {
+                if (getOrdersOpenHandler != null) {
+                    getOrdersOpenHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateGetTrades")) {
+                if (getTradesHandler != null) {
+                    getTradesHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateGetAccount")) {
+                if (getAccountHandler != null) {
+                    getAccountHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateGetBalance")) {
+                if(balanceHandler != null) {
+                    balanceHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateDepositAssets")) {
+                if(depositAssetsHandler != null) {
+                    depositAssetsHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateWithdrawAssets")) {
+                if(withdrawAssetsHandler != null) {
+                    withdrawAssetsHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateGetDepositHistory")) {
+                if(depositHistoryHandler != null) {
+                    depositHistoryHandler.handleMessage(response);
+                }
+            } else if(response.getString("action").equals("privateGetWithdrawalHistory")) {
+                if(withdrawalHistoryHandler != null) {
+                    withdrawalHistoryHandler.handleMessage(response);
                 }
             }
-            else if(response.getString("action").equals("getCandles")) {
-                if (this.candlesHandler != null) {
-                    this.candlesHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("getTicker24h")) {
-                if (this.ticker24hHandler != null) {
-                    this.ticker24hHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("getTickerPrice")) {
-                if (this.tickerPriceHandler != null) {
-                    this.tickerPriceHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("getTickerBook")) {
-                if (this.tickerBookHandler != null) {
-                    this.tickerBookHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateCreateOrder")) {
-                if (this.createOrderHandler != null) {
-                    this.createOrderHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateGetOrder")) {
-                if (this.getOrderHandler != null) {
-                    this.getOrderHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateUpdateOrder")) {
-                if (this.updateOrderHandler != null) {
-                    this.updateOrderHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateCancelOrder")) {
-                if (this.cancelOrderHandler != null) {
-                    this.cancelOrderHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateGetOrders")) {
-                if (this.getOrdersHandler != null) {
-                    this.getOrdersHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateCancelOrders")) {
-                if (this.cancelOrdersHandler != null) {
-                    this.cancelOrdersHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateGetOrdersOpen")) {
-                if (this.getOrdersOpenHandler != null) {
-                    this.getOrdersOpenHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateGetTrades")) {
-                if (this.getTradesHandler != null) {
-                    this.getTradesHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateGetAccount")) {
-                if (this.getAccountHandler != null) {
-                    this.getAccountHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateGetBalance")) {
-                if(this.balanceHandler != null) {
-                    this.balanceHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateDepositAssets")) {
-                if(this.depositAssetsHandler != null) {
-                    this.depositAssetsHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateWithdrawAssets")) {
-                if(this.withdrawAssetsHandler != null) {
-                    this.withdrawAssetsHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateGetDepositHistory")) {
-                if(this.depositHistoryHandler != null) {
-                    this.depositHistoryHandler.handleMessage(response);
-                }
-            }
-            else if(response.getString("action").equals("privateGetWithdrawalHistory")) {
-                if(this.withdrawalHistoryHandler != null) {
-                    this.withdrawalHistoryHandler.handleMessage(response);
-                }
-            }
-        }
-        else if (this.messageHandler != null) {
-            this.messageHandler.handleMessage(response);
+        } else if (messageHandler != null) {
+            messageHandler.handleMessage(response);
         }
     }
 
-    public void addMessageHandler(MessageHandler msgHandler) {
+    public void addMessageHandler(final MessageHandler msgHandler) {
         this.messageHandler = msgHandler;
     }
 
-    public void addErrorHandler(MessageHandler msgHandler) {
+    public void addErrorHandler(final MessageHandler msgHandler) {
         this.errorHandler = msgHandler;
     }
 
-    public void addTimeHandler(MessageHandler msgHandler) {
+    public void addTimeHandler(final MessageHandler msgHandler) {
         this.timeHandler = msgHandler;
     }
 
-    public void addMarketsHandler(MessageHandler msgHandler) {
+    public void addMarketsHandler(final MessageHandler msgHandler) {
         this.marketsHandler = msgHandler;
     }
 
-    public void addAssetsHandler(MessageHandler msgHandler) {
+    public void addAssetsHandler(final MessageHandler msgHandler) {
         this.assetsHandler = msgHandler;
     }
 
-    public void addBookHandler(MessageHandler msgHandler) {
+    public void addBookHandler(final MessageHandler msgHandler) {
         this.bookHandler = msgHandler;
     }
 
-    public void addTradesHandler(MessageHandler msgHandler) {
+    public void addTradesHandler(final MessageHandler msgHandler) {
         this.tradesHandler = msgHandler;
     }
 
-    public void addCandlesHandler(MessageHandler msgHandler) {
+    public void addCandlesHandler(final MessageHandler msgHandler) {
         this.candlesHandler = msgHandler;
     }
 
-    public void addTicker24hHandler(MessageHandler msgHandler) {
+    public void addTicker24hHandler(final MessageHandler msgHandler) {
         this.ticker24hHandler = msgHandler;
     }
 
-    public void addTickerPriceHandler(MessageHandler msgHandler) {
+    public void addTickerPriceHandler(final MessageHandler msgHandler) {
         this.tickerPriceHandler = msgHandler;
     }
 
-    public void addTickerBookHandler(MessageHandler msgHandler) {
+    public void addTickerBookHandler(final MessageHandler msgHandler) {
         this.tickerBookHandler = msgHandler;
     }
 
-    public void addPlaceOrderHandler(MessageHandler msgHandler) {
+    public void addPlaceOrderHandler(final MessageHandler msgHandler) {
         this.createOrderHandler = msgHandler;
     }
 
-    public void addGetOrderHandler(MessageHandler msgHandler) {
+    public void addGetOrderHandler(final MessageHandler msgHandler) {
         this.getOrderHandler = msgHandler;
     }
 
-    public void addUpdateOrderHandler(MessageHandler msgHandler) {
+    public void addUpdateOrderHandler(final MessageHandler msgHandler) {
         this.updateOrderHandler = msgHandler;
     }
 
-    public void addCancelOrderHandler(MessageHandler msgHandler) {
+    public void addCancelOrderHandler(final MessageHandler msgHandler) {
         this.cancelOrderHandler = msgHandler;
     }
 
-    public void addGetOrdersHandler(MessageHandler msgHandler) {
+    public void addGetOrdersHandler(final MessageHandler msgHandler) {
         this.getOrdersHandler = msgHandler;
     }
 
-    public void addCancelOrdersHandler(MessageHandler msgHandler) {
+    public void addCancelOrdersHandler(final MessageHandler msgHandler) {
         this.cancelOrdersHandler = msgHandler;
     }
 
-    public void addGetOrdersOpenHandler(MessageHandler msgHandler) {
+    public void addGetOrdersOpenHandler(final MessageHandler msgHandler) {
         this.getOrdersOpenHandler = msgHandler;
     }
 
-    public void addGetTradesHandler(MessageHandler msgHandler) {
+    public void addGetTradesHandler(final MessageHandler msgHandler) {
         this.getTradesHandler = msgHandler;
     }
 
-    public void addAccountHandler(MessageHandler msgHandler) {
+    public void addAccountHandler(final MessageHandler msgHandler) {
         this.getAccountHandler = msgHandler;
     }
 
-    public void addBalanceHandler(MessageHandler msgHandler) {
+    public void addBalanceHandler(final MessageHandler msgHandler) {
         this.balanceHandler = msgHandler;
     }
 
-    public void addDepositAssetsHandler(MessageHandler msgHandler) {
+    public void addDepositAssetsHandler(final MessageHandler msgHandler) {
         this.depositAssetsHandler = msgHandler;
     }
 
-    public void addWithdrawAssetsHandler(MessageHandler msgHandler) {
+    public void addWithdrawAssetsHandler(final MessageHandler msgHandler) {
         this.withdrawAssetsHandler = msgHandler;
     }
 
-    public void addDepositHistoryHandler(MessageHandler msgHandler) {
+    public void addDepositHistoryHandler(final MessageHandler msgHandler) {
         this.depositHistoryHandler = msgHandler;
     }
 
-    public void addWithdrawalHistoryHandler(MessageHandler msgHandler) {
+    public void addWithdrawalHistoryHandler(final MessageHandler msgHandler) {
         this.withdrawalHistoryHandler = msgHandler;
     }
 
-    public void addSubscriptionTickerHandler(String market, MessageHandler msgHandler) {
-        if(this.subscriptionTickerHandlerMap == null) {
-            this.subscriptionTickerHandlerMap = new HashMap<String, MessageHandler>();
+    public void addSubscriptionTickerHandler(final String market,final MessageHandler msgHandler) {
+        if(subscriptionTickerHandlerMap == null) {
+            subscriptionTickerHandlerMap = new HashMap<>();
         }
-        this.subscriptionTickerHandlerMap.put(market, msgHandler);
+        subscriptionTickerHandlerMap.put(market, msgHandler);
     }
 
-    public void copySubscriptionTickerHandler(HashMap<String, MessageHandler> map) {
+    public void copySubscriptionTickerHandler(final Map<String, MessageHandler> map) {
         this.subscriptionTickerHandlerMap = map;
     }
 
-    public void addSubscriptionTicker24hHandler(String market, MessageHandler msgHandler) {
-        if(this.subscriptionTicker24hHandlerMap == null) {
-            this.subscriptionTicker24hHandlerMap = new HashMap<String, MessageHandler>();
+    public void addSubscriptionTicker24hHandler(final String market,final MessageHandler msgHandler) {
+        if(subscriptionTicker24hHandlerMap == null) {
+            subscriptionTicker24hHandlerMap = new HashMap<>();
         }
-        this.subscriptionTicker24hHandlerMap.put(market, msgHandler);
+        subscriptionTicker24hHandlerMap.put(market, msgHandler);
     }
 
-    public void copySubscriptionTicker24hHandler(HashMap<String, MessageHandler> map) {
+    public void copySubscriptionTicker24hHandler(final Map<String, MessageHandler> map) {
         this.subscriptionTicker24hHandlerMap = map;
     }
 
-    public void addSubscriptionAccountHandler(String market, MessageHandler msgHandler) {
-        if(this.subscriptionAccountHandlerMap == null) {
-            this.subscriptionAccountHandlerMap = new HashMap<String, MessageHandler>();
+    public void addSubscriptionAccountHandler(final String market,final MessageHandler msgHandler) {
+        if(subscriptionAccountHandlerMap == null) {
+            subscriptionAccountHandlerMap = new HashMap<>();
         }
-        this.subscriptionAccountHandlerMap.put(market, msgHandler);
+        subscriptionAccountHandlerMap.put(market, msgHandler);
     }
 
-    public void copySubscriptionAccountHandler(HashMap<String, MessageHandler> map) {
+    public void copySubscriptionAccountHandler(final Map<String, MessageHandler> map) {
         this.subscriptionAccountHandlerMap = map;
     }
 
-    public void addSubscriptionCandlesHandler(String market, String interval, MessageHandler msgHandler) {
-        if(this.subscriptionCandlesHandlerMap == null) {
-            this.subscriptionCandlesHandlerMap = new HashMap<String, HashMap<String, MessageHandler>>();
+    public void addSubscriptionCandlesHandler(final String market,final String interval,final MessageHandler msgHandler) {
+        if (subscriptionCandlesHandlerMap == null) {
+            subscriptionCandlesHandlerMap = new HashMap<>();
         }
-        if(this.subscriptionCandlesHandlerMap.get(market) != null) {
-            this.subscriptionCandlesHandlerMap.get(market).put(interval, msgHandler);
+        if (subscriptionCandlesHandlerMap.get(market) != null) {
+            subscriptionCandlesHandlerMap.get(market).put(interval, msgHandler);
         } else {
-            this.subscriptionCandlesHandlerMap.put(market, new HashMap<String, MessageHandler>());
-            this.subscriptionCandlesHandlerMap.get(market).put(interval, msgHandler);
+            subscriptionCandlesHandlerMap.put(market, new HashMap<>());
+            subscriptionCandlesHandlerMap.get(market).put(interval, msgHandler);
         }
     }
 
-    public void copySubscriptionCandlesHandler(HashMap<String, HashMap<String, MessageHandler>> map) {
+    public void copySubscriptionCandlesHandler(final Map<String, HashMap<String, MessageHandler>> map) {
         this.subscriptionCandlesHandlerMap = map;
     }
 
-    public void addSubscriptionTradesHandler(String market, MessageHandler msgHandler) {
-        if(this.subscriptionTradesHandlerMap == null) {
-            this.subscriptionTradesHandlerMap = new HashMap<String, MessageHandler>();
+    public void addSubscriptionTradesHandler(final String market,final MessageHandler msgHandler) {
+        if (subscriptionTradesHandlerMap == null) {
+            subscriptionTradesHandlerMap = new HashMap<>();
         }
-        this.subscriptionTradesHandlerMap.put(market, msgHandler);
+        subscriptionTradesHandlerMap.put(market, msgHandler);
     }
 
-    public void copySubscriptionTradesHandler(HashMap<String, MessageHandler> map) {
+    public void copySubscriptionTradesHandler(final Map<String, MessageHandler> map) {
         this.subscriptionTradesHandlerMap = map;
     }
 
-    public void addSubscriptionBookUpdateHandler(String market, MessageHandler msgHandler) {
-        if(this.subscriptionBookUpdateHandlerMap == null) {
-            this.subscriptionBookUpdateHandlerMap = new HashMap<String, MessageHandler>();
+    public void addSubscriptionBookUpdateHandler(final String market,final MessageHandler msgHandler) {
+        if (subscriptionBookUpdateHandlerMap == null) {
+            subscriptionBookUpdateHandlerMap = new HashMap<>();
         }
-        this.subscriptionBookUpdateHandlerMap.put(market, msgHandler);
+        subscriptionBookUpdateHandlerMap.put(market, msgHandler);
     }
 
-    public void copySubscriptionBookUpdateHandler(HashMap<String, MessageHandler> map) {
+    public void copySubscriptionBookUpdateHandler(final Map<String, MessageHandler> map) {
         this.subscriptionBookUpdateHandlerMap = map;
     }
 
-    public void addAuthenticateHandler(MessageHandler msgHandler) {
+    public void addAuthenticateHandler(final MessageHandler msgHandler) {
         this.authenticateHandler = msgHandler;
     }
 
-    public void addSubscriptionBookHandler(String market, BookHandler msgHandler) {
-        if(this.subscriptionBookHandlerMap == null) {
-            this.subscriptionBookHandlerMap = new HashMap<String, BookHandler>();
+    public void addSubscriptionBookHandler(final String market,final BookHandler msgHandler) {
+        if (subscriptionBookHandlerMap == null) {
+            subscriptionBookHandlerMap = new HashMap<>();
         }
-        this.subscriptionBookHandlerMap.put(market, msgHandler);
+        subscriptionBookHandlerMap.put(market, msgHandler);
     }
 
-    public void copySubscriptionBookHandler(HashMap<String, BookHandler> map) {
+    public void copySubscriptionBookHandler(final Map<String, BookHandler> map) {
         this.subscriptionBookHandlerMap = map;
     }
 
-    public void sendMessage(String message) {
-        bitvavo.debugToConsole("Sending message " + message);
-        this.userSession.getAsyncRemote().sendText(message);
+    public void sendMessage(final String message) {
+        debug("Sending message " + message);
+        userSession.getAsyncRemote().sendText(message);
     }
 
     public static interface MessageHandler {
@@ -784,5 +753,9 @@ public class WebsocketClientEndpoint {
 
     public static interface BookHandler {
         public void handleBook(Map<String, Object> book);
+    }
+
+    public void setKeepBookCopy(final boolean keepBookCopy) {
+        this.keepBookCopy = keepBookCopy;
     }
 }
